+++ legacy/admin/auth_verify.php
@@ -11,18 +11,23 @@ function is_session_expired()
     }
 }
 
-if (isset($_REQUEST["service_body_id"])) {
+if (isset($_REQUEST["service_body_id"]) && $_REQUEST["service_body_id"] > -1) {
     $found = false;
-    foreach (getServiceBodiesRights() as $service_body) {
-        if ($service_body->id == $_REQUEST['service_body_id']) {
-            $found = true;
-            continue;
+    $service_body_rights = getServiceBodiesRights();
+    if (isset($service_body_rights)) {
+        foreach ($service_body_rights as $service_body) {
+            if ($service_body->id == $_REQUEST['service_body_id']) {
+                $found = true;
+                continue;
+            }
         }
     }
 
     if (!$found) {
         session_unset();
     }
+} elseif (isset($_REQUEST["service_body_id"]) && $_REQUEST["service_body_id"] == -1) {
+    $found = true;
 }
 
 $expired = is_session_expired();