+++ legacy/admin/cdr_api.php
@@ -1,4 +1,4 @@
 <?php
-require_once __DIR__ . '/../_includes/functions.php';
+require_once 'auth_verify.php';
 header("content-type: application/json");
 echo json_encode(adjustedCallRecords(getReportsServiceBodies(), $_REQUEST['date_range_start'], $_REQUEST['date_range_end']));