+++ RELEASENOTES.md
@@ -6,6 +6,8 @@
 * Added support for customizing SMS for incoming calls. [#636]
 * Show missed calls also as a percentage of total calls received.
 * Metric endpoint now returns volunteer missed and answered metrics, not exposed to UI yet. [#632]
+* Fixed issue when metric summaries wouldn't zero out between switching service bodies.
+* Fixed issue with metric summary for Volunteer SMS which wasn't reporting correctly.
 * Fixed issues with unassigned aka "General" metrics, when no service body is known or defined at a point in a call or SMS. [#627]
 * Fixed an issue with reports not recursing more than one level. [#620]